Move files to trash, might be bugs
(emms-browser-move-files-to-trash) mimics the behaviour
of (emms-browser-delete-files), using (move-file-to-trash FILENAME).
This allows users to keep the `delete' functionality of `delete-file`
and `delete-directory` (which can be set to send to trash instead), and
also gives some confidence in the music removal and organisation
features of emms (that they won't accidentally permanently delete
music), and have full use of these features, including removing covers
and parents.

Added hook `emms-browser-move-files-to-trash-hook` and
function (de-trash-covers-and-parents) is called with that
hook.  (de-trash-covers-and-parents)
mimics (de-kill-covers-and-parents), which is found in commentary of
emms-browser.el.

I have not fully tested the new functionality.